do-flow-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
httpx==0.25.2
aiofiles==23.2.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2
openai==1.3.8
langchain==0.0.338
langchain-openai==0.0.2
streamlit==1.28.1
prometheus-client==0.19.0
structlog==23.2.0
cryptography==41.0.7
email-validator==2.1.0
Jinja2==3.1.2
python-dateutil==2.8.2
requests==2.31.0
websockets==12.0
asyncpg==0.29.0
pyjwt==2.8.0
bcrypt==4.1.1
pytz==2023.3
schedule==1.2.0
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
Pillow==10.1.0
qrcode==7.4.2
stripe==7.8.0
twilio==8.10.3
sendgrid==6.10.0
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0